Actually rom.group is pretty useless
